+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/Events/CallAutomationEventParser.cs
@@ -120,6 +120,10 @@ namespace Azure.Communication.CallAutomation
                     return RecognizeFailed.Deserialize(eventData);
                 case nameof(RecognizeCanceled):
                     return RecognizeCanceled.Deserialize(eventData);
+                case nameof(RemoveParticipantSucceeded):
+                    return RemoveParticipantSucceeded.Deserialize(eventData);
+                case nameof(RemoveParticipantFailed):
+                    return RemoveParticipantFailed.Deserialize(eventData);
                 default:
                     return null;
             }