+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/Events/CollectTonesResult.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
+using Azure.Communication.CallAutomation.Models;
 using System.Linq;
 using Azure.Core;
 
@@ -15,6 +16,11 @@ namespace Azure.Communication.CallAutomation
         /// </summary>
         [CodeGenMember("Tones")]
         public IReadOnlyList<DtmfTone> Tones { get; }
+
+        /// <summary>
+        /// The RecognizeResultType of this RecognizeResult.
+        /// </summary>
+        public override RecognizeResultType ResultType => RecognizeResultType.CollectTonesResult;
 
         /// <summary>
         /// Convert the collection of tones to a string like "12345#".