+++ sdk/healthinsights/Azure.Health.Insights.ClinicalMatching/src/Constants.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Health.Insights.ClinicalMatching
+{
+    internal static class Constants
+    {
+        public const string AuthorizationHeader = "Ocp-Apim-Subscription-Key";
+
+        public const string DefaultCognitiveScope = "https://cognitiveservices.azure.com/.default";
+    }
+}