+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Bugs Fixed
 
+- Set the Activity status to `Error` on failed activity source activities.
+- Mark the `Azure.Core.Http.Request` span as failed if the request fails with an exception thrown in the pipeline.
+- Fixed equality comparison when comparing a `string` to a `ContentType` instance.
+
 ### Other Changes
 
 ## 1.30.0 (2023-03-09)